move lists into /component/list
