--- src/component/SettingsPanel.jsx
@@ -1,8 +1,0 @@
-import React from 'react'
-
-export default function SettingsPanel({display}) {
-    if(!display) return null
-    return(
-        <div>SettingsPanel</div>
-    )
-}
